Put utils.py into common/model_utils.py
